extracted the token functionality to its own module
